requirements.txt
# AI/ML libraries (install first to avoid conflicts)
face-recognition==1.3.0
dlib==20.0.0
mediapipe==0.10.9
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.2.0

# FastAPI and web server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# LangGraph and LangChain
langgraph==0.2.28
langchain-core==0.3.15
langchain-google-genai==2.0.4

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-vision==3.7.0
google-cloud-speech>=2.34.0  # V2 API with Chirp 3 support (includes V1 for backward compatibility)
google-generativeai==0.8.3

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
requests==2.31.0
psutil==5.9.8

# Monitoring (optional)
prometheus-client==0.20.0

